refactor farm and router tests
