+++ dex/farm/tests/farm_multi_user_test.rs
@@ -2,12 +2,13 @@ use common_structs::FarmTokenAttributes;
 use elrond_wasm_debug::{managed_biguint, rust_biguint, DebugApi};
 
 pub mod farm_setup;
-use farm_setup::*;
+use farm_setup::multi_user_farm_setup::*;
 
 #[test]
 fn farm_with_no_boost_test() {
     let _ = DebugApi::dummy();
-    let mut farm_setup = FarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
+    let mut farm_setup =
+        MultiUserFarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
 
     // first user enter farm
     let first_farm_token_amount = 100_000_000;
@@ -99,7 +100,8 @@ fn farm_with_no_boost_test() {
 #[test]
 fn farm_with_boosted_yields_test() {
     let _ = DebugApi::dummy();
-    let mut farm_setup = FarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
+    let mut farm_setup =
+        MultiUserFarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
 
     farm_setup.set_boosted_yields_rewards_percentage(BOOSTED_YIELDS_PERCENTAGE);
     farm_setup.set_boosted_yields_factors();
@@ -207,7 +209,8 @@ fn farm_with_boosted_yields_test() {
 #[test]
 fn farm_known_proxy_test() {
     let _ = DebugApi::dummy();
-    let mut farm_setup = FarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
+    let mut farm_setup =
+        MultiUserFarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
 
     // first user enter farm
     let first_farm_token_amount = 100_000_000;
@@ -300,7 +303,8 @@ fn farm_known_proxy_test() {
 #[test]
 fn farm_multiple_claim_weeks_with_collect_undistributed_rewards_test() {
     let _ = DebugApi::dummy();
-    let mut farm_setup = FarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
+    let mut farm_setup =
+        MultiUserFarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
 
     farm_setup.set_boosted_yields_rewards_percentage(BOOSTED_YIELDS_PERCENTAGE);
     farm_setup.set_boosted_yields_factors();
@@ -519,7 +523,8 @@ fn farm_multiple_claim_weeks_with_collect_undistributed_rewards_test() {
 #[test]
 fn farm_enter_with_multiple_farm_token() {
     let _ = DebugApi::dummy();
-    let mut farm_setup = FarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
+    let mut farm_setup =
+        MultiUserFarmSetup::new(farm::contract_obj, energy_factory_mock::contract_obj);
 
     farm_setup.set_boosted_yields_rewards_percentage(BOOSTED_YIELDS_PERCENTAGE);
     farm_setup.set_boosted_yields_factors();
