+++ locked-asset/energy-factory/tests/energy_factory_setup/mod.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use elrond_wasm::{
     elrond_codec::multi_types::OptionalValue,
     storage::mappers::StorageTokenWrapper,
